Save and Load Fix

+++ Assets/2D-ARPG/Scripts/PlayerScripts/HpMpRegen.cs
@@ -35,7 +35,7 @@ public class HpMpRegen : MonoBehaviour {
 	}
 	
 	void HPRecovery(){
-		float amount = stat.totalStat.health * hpRegen / 100;
+        int amount = stat.totalStat.health * hpRegen / 100;
 		if(amount <= 1){
 			amount = 1;
 		}
